+++ lib/src/opacity.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pressable/pressable.dart';
 import 'package:pressable/src/base.dart';
+import 'package:pressable/src/provider.dart';
 
 class PressableOpacity extends Pressable {
   const PressableOpacity({
@@ -14,13 +15,19 @@ class PressableOpacity extends Pressable {
   final Widget child;
   final VoidCallback? onPressed;
   final VoidCallback? onLongPressed;
-  final PressableOpacityTheme theme;
+  final PressableOpacityTheme? theme;
 
   @override
   _PressableOpacityState createState() => _PressableOpacityState();
 }
 
 class _PressableOpacityState extends PressableBaseState<PressableOpacity> {
+  PressableOpacityTheme get theme {
+    return widget.theme ??
+        DefaultPressableTheme.of(context)?.opacityTheme ??
+        const PressableOpacityTheme();
+  }
+
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
@@ -31,9 +38,9 @@ class _PressableOpacityState extends PressableBaseState<PressableOpacity> {
       onLongPress: widget.onLongPressed,
       behavior: HitTestBehavior.opaque,
       child: AnimatedOpacity(
-        opacity: isPressed ? widget.theme.opacityFactor : 1.0,
-        duration: widget.theme.duration,
-        curve: widget.theme.curve,
+        opacity: isPressed ? theme.opacityFactor : 1.0,
+        duration: theme.duration,
+        curve: theme.curve,
         child: widget.child,
       ),
     );